Create Perfect Number
A perfect number program in c is a number, whose sum is equal to its positive divisor but excludes the number itself. For instance, the number 6 is divisible by 1, 2, and 3 completely. The number also divides itself but for a perfect number, the number divisible by itself is excluded. When the divisors are added, 1+2+3 the result is also 6 which is equal to the number that is considered. Hence the number 6 is perfect.